# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Invoice API",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/products": {
            "get": {
                "tags": [
                    "Products"
                ],
                "summary": "Retrieve a paginated list of products",
                "description": "This endpoint returns a paginated list of products available in the system.",
                "operationId": "1bfaa78d1c2c3848ab8165c5dadcad3e",
                "responses": {
                    "200": {
                        "description": "Paginated list of products retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "current_page": {
                                            "description": "Current page number",
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "data": {
                                            "description": "List of products for the current page",
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "description": "Product ID",
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "name": {
                                                        "description": "Product name",
                                                        "type": "string",
                                                        "example": "Sample Product"
                                                    },
                                                    "description": {
                                                        "description": "Product description",
                                                        "type": "string",
                                                        "example": "This is a sample product."
                                                    },
                                                    "price": {
                                                        "description": "Product price",
                                                        "type": "number",
                                                        "format": "float",
                                                        "example": 199.99
                                                    },
                                                    "quantity": {
                                                        "description": "Quantity in stock",
                                                        "type": "integer",
                                                        "example": 100
                                                    },
                                                    "category_id": {
                                                        "description": "ID of the associated category",
                                                        "type": "integer",
                                                        "example": 2
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        },
                                        "first_page_url": {
                                            "description": "URL for the first page",
                                            "type": "string",
                                            "example": "http://example.com/api/products?page=1"
                                        },
                                        "from": {
                                            "description": "Starting record number on the current page",
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "last_page": {
                                            "description": "Total number of pages",
                                            "type": "integer",
                                            "example": 10
                                        },
                                        "last_page_url": {
                                            "description": "URL for the last page",
                                            "type": "string",
                                            "example": "http://example.com/api/products?page=10"
                                        },
                                        "next_page_url": {
                                            "description": "URL for the next page, if available",
                                            "type": "string",
                                            "example": "http://example.com/api/products?page=2",
                                            "nullable": true
                                        },
                                        "path": {
                                            "description": "Base URL for the API",
                                            "type": "string",
                                            "example": "http://example.com/api/products"
                                        },
                                        "per_page": {
                                            "description": "Number of records per page",
                                            "type": "integer",
                                            "example": 15
                                        },
                                        "prev_page_url": {
                                            "description": "URL for the previous page, if available",
                                            "type": "string",
                                            "example": null,
                                            "nullable": true
                                        },
                                        "to": {
                                            "description": "Ending record number on the current page",
                                            "type": "integer",
                                            "example": 15
                                        },
                                        "total": {
                                            "description": "Total number of records available",
                                            "type": "integer",
                                            "example": 150
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Server Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Server Error"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "An unexpected error occurred."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/product/save": {
            "post": {
                "tags": [
                    "Products"
                ],
                "summary": "Add a new product or update an existing product",
                "description": "This endpoint is used to create a new product or update an existing product based on the provided product ID.",
                "operationId": "a9f8f29d8ec492bc4385e6a4d8406543",
                "requestBody": {
                    "description": "Product data to create or update",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "description",
                                    "price",
                                    "quantity",
                                    "category_id"
                                ],
                                "properties": {
                                    "id": {
                                        "description": "Product ID (for update)",
                                        "type": "integer",
                                        "example": null
                                    },
                                    "name": {
                                        "description": "Product name",
                                        "type": "string",
                                        "example": "Sample Product"
                                    },
                                    "description": {
                                        "description": "Product description",
                                        "type": "string",
                                        "example": "This is a sample product."
                                    },
                                    "price": {
                                        "description": "Product price",
                                        "type": "number",
                                        "format": "float",
                                        "example": 199.99
                                    },
                                    "quantity": {
                                        "description": "Quantity in stock",
                                        "type": "integer",
                                        "example": 100
                                    },
                                    "category_id": {
                                        "description": "ID of the product category",
                                        "type": "integer",
                                        "example": 2
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Product created or updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "description": "Product ID",
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "name": {
                                            "type": "string",
                                            "example": "Sample Product"
                                        },
                                        "description": {
                                            "type": "string",
                                            "example": "This is a sample product."
                                        },
                                        "price": {
                                            "type": "number",
                                            "format": "float",
                                            "example": 199.99
                                        },
                                        "quantity": {
                                            "type": "integer",
                                            "example": 100
                                        },
                                        "category_id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "created_at": {
                                            "type": "string",
                                            "format": "date-time",
                                            "example": "2023-11-13T10:00:00Z"
                                        },
                                        "updated_at": {
                                            "type": "string",
                                            "format": "date-time",
                                            "example": "2023-11-13T10:00:00Z"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request - Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Bad Request"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Validation failed for the provided product data."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Server Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Server Error"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/product/view/{id}": {
            "get": {
                "tags": [
                    "Products"
                ],
                "summary": "Get a product by ID",
                "description": "Retrieve a product's details by its ID.",
                "operationId": "546821070befb8b59e7d52bd434f0551",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the product to retrieve",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Product details retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "name": {
                                            "type": "string",
                                            "example": "Sample Product"
                                        },
                                        "description": {
                                            "type": "string",
                                            "example": "This is a sample product."
                                        },
                                        "price": {
                                            "type": "number",
                                            "format": "float",
                                            "example": 199.99
                                        },
                                        "quantity": {
                                            "type": "integer",
                                            "example": 100
                                        },
                                        "category_id": {
                                            "type": "integer",
                                            "example": 2
                                        },
                                        "created_at": {
                                            "type": "string",
                                            "format": "date-time",
                                            "example": "2023-10-04T06:13:12.000000Z"
                                        },
                                        "updated_at": {
                                            "type": "string",
                                            "format": "date-time",
                                            "example": "2023-10-04T06:13:12.000000Z"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Product not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Product not found"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "No product found with the given ID."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
...
}